+++ apps/executor/CHANGELOG.md
@@ -1,5 +1,17 @@
 # executor
 
+## 1.3.0-beta.5
+
+### Patch Changes
+
+- 0316b25: Compact openapi sources
+
+## 1.3.0-beta.4
+
+### Patch Changes
+
+- 787dd66: Improve Google Discovery sources view
+
 ## 1.3.0-beta.3
 
 ### Patch Changes